API/https files moved to new dir
