+++ lib/pages/login_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 
+import '../apis/auth_client.dart';
 import '../components/license.dart';
 import '../configs/configs.dart' as configs;
 import '../constants/strings.dart';
 import '../pages/sign_up_page.dart';
 import '../pages/view_source_code.dart';
-import '../utils/auth_client.dart';
 
 class LogInPage extends StatelessWidget {
